test: Split apply_stack and delete_stack tests
The single `apply_delete_stack.rs` integration test has been split into
separate `apply_stack.rs` and `delete_stack.rs` tests. A new `common`
module has been introduced with the functions and constants that are
used in both, or are likely to be used by other tests in future.
