+++ tests/apply_stack.rs
@@ -1,50 +1,15 @@
-use std::{env, time::Duration};
+pub mod common;
 
 use futures_util::StreamExt;
-use rusoto_cloudformation::{CloudFormationClient, DescribeStacksInput};
-use rusoto_core::HttpClient;
-use rusoto_credential::{AutoRefreshingProvider, ChainProvider};
 
 use cloudformatious::{
     change_set::{Action, ExecutionStatus},
-    ApplyStackError, ApplyStackInput, ChangeSetStatus, CloudFormatious, DeleteStackInput,
-    Parameter, ResourceStatus, StackFailure, StackStatus, TemplateSource,
+    ApplyStackError, ApplyStackInput, ChangeSetStatus, CloudFormatious, ResourceStatus,
+    StackFailure, StackStatus, TemplateSource,
 };
 
-const NAME_PREFIX: &str = "cloudformatious-testing-";
-const EMPTY_TEMPLATE: &str = r#"{
-    "Conditions": {
-        "Never": { "Fn::Equals": [true, false] }
-    },
-    "Resources": {
-        "Fake": {
-            "Type": "Custom::Fake",
-            "Condition": Never
-        }
-    }
-}"#;
-const NON_EMPTY_TEMPLATE: &str = r#"{
-    "Parameters": {
-        "CidrBlock": {
-            "Type": "String"
-        }
-    },
-    "Resources": {
-        "Subnet": {
-            "Type": "AWS::EC2::Subnet",
-            "Properties": {
-                "CidrBlock": {"Ref": "CidrBlock"},
-                "Tags": [
-                    {
-                        "Key": "VpcId",
-                        "Value": {"Fn::ImportValue": "cloudformatious-testing-VpcId"}
-                    }
-                ],
-                "VpcId": {"Fn::ImportValue": "cloudformatious-testing-VpcId"}
-            }
-        }
-    }
-}"#;
+use common::{clean_up, generated_name, get_client, EMPTY_TEMPLATE};
+
 const FAILING_TEMPLATE: &str = r#"
             {
                 "Resources": {
@@ -441,168 +406,4 @@ async fn update_stack_fut_err() -> Result<(), Box<dyn std::error::Error>> {
     clean_up(&client, stack_name).await?;
 
     Ok(())
-}
-
-#[tokio::test]
-async fn delete_stack_fut_ok() -> Result<(), Box<dyn std::error::Error>> {
-    let client = get_client();
-
-    let stack_name = generated_name();
-    let input = ApplyStackInput::new(&stack_name, TemplateSource::inline(EMPTY_TEMPLATE));
-    let stack = client.apply_stack(input).await?;
-
-    let input = DeleteStackInput::new(&stack_name);
-    client.delete_stack(input).await?;
-
-    let input = DescribeStacksInput {
-        stack_name: Some(stack.stack_id),
-        ..DescribeStacksInput::default()
-    };
-    let stack = {
-        use rusoto_cloudformation::CloudFormation;
-        client
-            .describe_stacks(input)
-            .await?
-            .stacks
-            .unwrap()
-            .pop()
-            .unwrap()
-    };
-
-    assert_eq!(stack.stack_status, "DELETE_COMPLETE");
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn delete_stack_stream_ok() -> Result<(), Box<dyn std::error::Error>> {
-    let client = get_client();
-
-    let stack_name = generated_name();
-    let input = ApplyStackInput::new(&stack_name, TemplateSource::inline(NON_EMPTY_TEMPLATE))
-        .set_parameters([Parameter {
-            key: "CidrBlock".to_string(),
-            value: "10.0.0.0/28".to_string(),
-        }]);
-    let stack = client.apply_stack(input).await?;
-
-    let input = DeleteStackInput::new(&stack_name);
-    let mut delete = client.delete_stack(input);
-
-    let events: Vec<_> = delete
-        .events()
-        .map(|event| {
-            (
-                event.logical_resource_id().to_string(),
-                event.resource_status().to_string(),
-            )
-        })
-        .collect()
-        .await;
-    delete.await?;
-
-    assert_eq!(
-        events,
-        vec![
-            (stack_name.clone(), "DELETE_IN_PROGRESS".to_string()),
-            ("Subnet".to_string(), "DELETE_IN_PROGRESS".to_string()),
-            ("Subnet".to_string(), "DELETE_COMPLETE".to_string()),
-            (stack_name.clone(), "DELETE_COMPLETE".to_string())
-        ]
-    );
-
-    let input = DescribeStacksInput {
-        stack_name: Some(stack.stack_id),
-        ..DescribeStacksInput::default()
-    };
-    let stack = {
-        use rusoto_cloudformation::CloudFormation;
-        client
-            .describe_stacks(input)
-            .await?
-            .stacks
-            .unwrap()
-            .pop()
-            .unwrap()
-    };
-
-    assert_eq!(stack.stack_status, "DELETE_COMPLETE");
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn delete_stack_fut_noop() -> Result<(), Box<dyn std::error::Error>> {
-    let client = get_client();
-
-    let stack_name = generated_name();
-    let input = DeleteStackInput::new(&stack_name);
-    client.delete_stack(input).await?;
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn delete_stack_stream_noop() -> Result<(), Box<dyn std::error::Error>> {
-    let client = get_client();
-
-    let stack_name = generated_name();
-    let input = DeleteStackInput::new(&stack_name);
-    let mut delete = client.delete_stack(input);
-
-    let events: Vec<_> = delete.events().collect().await;
-    delete.await?;
-
-    assert_eq!(events, vec![]);
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn delete_stack_idempotent() -> Result<(), Box<dyn std::error::Error>> {
-    let client = get_client();
-
-    let stack_name = generated_name();
-    let input = ApplyStackInput::new(&stack_name, TemplateSource::inline(EMPTY_TEMPLATE));
-    let stack = client.apply_stack(input).await?;
-
-    let input = DeleteStackInput::new(&stack.stack_id);
-    client.delete_stack(input).await?;
-
-    let input = DeleteStackInput::new(&stack.stack_id);
-    client.delete_stack(input).await?;
-
-    Ok(())
-}
-
-fn get_client() -> CloudFormationClient {
-    let client = HttpClient::new().unwrap();
-
-    let mut credentials = AutoRefreshingProvider::new(ChainProvider::new()).unwrap();
-    credentials.get_mut().set_timeout(Duration::from_secs(1));
-
-    let region = env::var("AWS_REGION").expect("You must set AWS_REGION to run these tests");
-    let region = region.parse().expect("Invalid AWS region");
-
-    CloudFormationClient::new_with(client, credentials, region)
-}
-
-fn generated_name() -> String {
-    format!("{}{}", NAME_PREFIX, fastrand::u32(..))
-}
-
-async fn clean_up(
-    client: &CloudFormationClient,
-    stack_name: String,
-) -> Result<(), Box<dyn std::error::Error>> {
-    use rusoto_cloudformation::{CloudFormation, DeleteStackInput};
-    CloudFormation::delete_stack(
-        client,
-        DeleteStackInput {
-            stack_name,
-            ..DeleteStackInput::default()
-        },
-    )
-    .await
-    .map_err(Into::into)
 }
